move boilerplate FunctionN into akka.japi.function package
This is a temporary solution until the classes are added to akka-actor.
