+++ akka-stream-tests/src/test/java/akka/stream/javadsl/BidiFlowTest.java
@@ -20,7 +20,6 @@ import akka.japi.Pair;
 import akka.stream.*;
 import akka.stream.testkit.AkkaSpec;
 import akka.stream.javadsl.FlowGraph.Builder;
-import akka.stream.javadsl.japi.*;
 import akka.japi.function.*;
 import akka.util.ByteString;
 import static org.junit.Assert.assertEquals;
