+++ akka-stream/src/main/boilerplate/akka/stream/javadsl/GraphCreate.scala.template
@@ -5,6 +5,7 @@ package akka.stream.javadsl
 
 import akka.stream.scaladsl
 import akka.stream.{ Inlet, Shape, Graph }
+import akka.japi.function
 
 trait GraphCreate {
 
@@ -17,7 +18,7 @@ trait GraphCreate {
    * The created graph must have all ports connected, otherwise an [[IllegalArgumentException]] is thrown.
    */
   @throws(classOf[IllegalArgumentException])
-  def closed(block: akka.japi.function.Procedure[FlowGraph.Builder[Unit]]): RunnableFlow[Unit] =
+  def closed(block: function.Procedure[FlowGraph.Builder[Unit]]): RunnableFlow[Unit] =
     scaladsl.FlowGraph.closed() { b ⇒ block.apply(b.asJava) }
 
   /**
@@ -25,7 +26,7 @@ trait GraphCreate {
    *
    * Partial graphs are allowed to have unconnected ports.
    */
-  def partial[S <: Shape](block: akka.japi.function.Function[FlowGraph.Builder[Unit], S]): Graph[S, Unit] =
+  def partial[S <: Shape](block: function.Function[FlowGraph.Builder[Unit], S]): Graph[S, Unit] =
     scaladsl.FlowGraph.partial() { b ⇒ block.apply(b.asJava) }
 
   /**
@@ -36,7 +37,7 @@ trait GraphCreate {
    */
   @throws(classOf[IllegalArgumentException])
   def closed[S1 <: Shape, M](g1: Graph[S1, M],
-      block: japi.Procedure2[FlowGraph.Builder[M], S1]): RunnableFlow[M] =
+      block: function.Procedure2[FlowGraph.Builder[M], S1]): RunnableFlow[M] =
     scaladsl.FlowGraph.closed(g1) { b ⇒ s => block.apply(b.asJava, s) }
 
   /**
@@ -46,7 +47,7 @@ trait GraphCreate {
    * Partial graphs are allowed to have unconnected ports.
    */
   def partial[S1 <: Shape, S <: Shape, M](g1: Graph[S1, M],
-      block: akka.japi.function.Function2[FlowGraph.Builder[M], S1, S]): Graph[S, M] =
+      block: function.Function2[FlowGraph.Builder[M], S1, S]): Graph[S, M] =
     scaladsl.FlowGraph.partial(g1) { b ⇒ s => block.apply(b.asJava, s) }
 
   /**
@@ -56,8 +57,8 @@ trait GraphCreate {
    * The created graph must have all ports connected, otherwise an [[IllegalArgumentException]] is thrown.
    */
   @throws(classOf[IllegalArgumentException])
-  def closed[S1 <: Shape, S2 <: Shape, M1, M2, M](g1: Graph[S1, M1], g2: Graph[S2, M2], combineMat: akka.japi.function.Function2[M1, M2, M],
-      block: japi.Procedure3[FlowGraph.Builder[M], S1, S2]): RunnableFlow[M] =
+  def closed[S1 <: Shape, S2 <: Shape, M1, M2, M](g1: Graph[S1, M1], g2: Graph[S2, M2], combineMat: function.Function2[M1, M2, M],
+      block: function.Procedure3[FlowGraph.Builder[M], S1, S2]): RunnableFlow[M] =
     scaladsl.FlowGraph.closed(g1, g2)(combineMat.apply _) { b => (s1, s2) => block.apply(b.asJava, s1, s2) }
 
   /**
@@ -66,8 +67,8 @@ trait GraphCreate {
    *
    * Partial graphs are allowed to have unconnected ports.
    */
-  def partial[S1 <: Shape, S2 <: Shape, S <: Shape, M1, M2, M](g1: Graph[S1, M1], g2: Graph[S2, M2], combineMat: akka.japi.function.Function2[M1, M2, M],
-      block: japi.Function3[FlowGraph.Builder[M], S1, S2, S]): Graph[S, M] =
+  def partial[S1 <: Shape, S2 <: Shape, S <: Shape, M1, M2, M](g1: Graph[S1, M1], g2: Graph[S2, M2], combineMat: function.Function2[M1, M2, M],
+      block: function.Function3[FlowGraph.Builder[M], S1, S2, S]): Graph[S, M] =
     scaladsl.FlowGraph.partial(g1, g2)(combineMat.apply _) { b => (s1, s2) => block.apply(b.asJava, s1, s2) }
 
   [3..21#/**
@@ -77,8 +78,8 @@ trait GraphCreate {
    * The created graph must have all ports connected, otherwise an [[IllegalArgumentException]] is thrown.
    */
   @throws(classOf[IllegalArgumentException])
-  def closed1[[#S1 <: Shape#], [#M1#], M]([#g1: Graph[S1, M1]#], combineMat: japi.Function1[[#M1#], M],
-      block: japi.Procedure2[FlowGraph.Builder[M], [#S1#]]): RunnableFlow[M] =
+  def closed1[[#S1 <: Shape#], [#M1#], M]([#g1: Graph[S1, M1]#], combineMat: function.Function1[[#M1#], M],
+      block: function.Procedure2[FlowGraph.Builder[M], [#S1#]]): RunnableFlow[M] =
     scaladsl.FlowGraph.closed([#g1#])(combineMat.apply _) { b => ([#s1#]) => block.apply(b.asJava, [#s1#]) }
 
   /**
@@ -87,8 +88,8 @@ trait GraphCreate {
    *
    * Partial graphs are allowed to have unconnected ports.
    */
-  def partial1[[#S1 <: Shape#], S <: Shape, [#M1#], M]([#g1: Graph[S1, M1]#], combineMat: japi.Function1[[#M1#], M],
-      block: japi.Function2[FlowGraph.Builder[M], [#S1#], S]): Graph[S, M] =
+  def partial1[[#S1 <: Shape#], S <: Shape, [#M1#], M]([#g1: Graph[S1, M1]#], combineMat: function.Function1[[#M1#], M],
+      block: function.Function2[FlowGraph.Builder[M], [#S1#], S]): Graph[S, M] =
     scaladsl.FlowGraph.partial([#g1#])(combineMat.apply _) { b => ([#s1#]) => block.apply(b.asJava, [#s1#]) }#
   
   ]
