+++ akka-stream/src/main/boilerplate/akka/stream/javadsl/ZipWith.scala.template
@@ -5,6 +5,7 @@ package akka.stream.javadsl
 
 import akka.stream._
 import akka.stream.scaladsl
+import akka.japi.function
 
 /**
  * Combine the elements of multiple streams into a stream of combined elements using a combiner function.
@@ -25,7 +26,7 @@ object ZipWith {
    * @param f zipping-function from the input values to the output value
    * @param attributes optional attributes for this vertex
    */
-  def create[A, B, Out](f: akka.japi.function.Function2[A, B, Out]): Graph[FanInShape2[A, B, Out], Unit] =
+  def create[A, B, Out](f: function.Function2[A, B, Out]): Graph[FanInShape2[A, B, Out], Unit] =
     scaladsl.ZipWith(f.apply _)
 
   [3..20#/** Create a new `ZipWith` specialized for 1 inputs.
@@ -33,7 +34,7 @@ object ZipWith {
    * @param f zipping-function from the input values to the output value
    * @param attributes optional attributes for this vertex
    */
-  def create1[[#T1#], Out](f: japi.Function1[[#T1#], Out]): Graph[FanInShape1[[#T1#], Out], Unit] =
+  def create1[[#T1#], Out](f: function.Function1[[#T1#], Out]): Graph[FanInShape1[[#T1#], Out], Unit] =
     scaladsl.ZipWith(f.apply _)#
     
   ]
